Update index.html
Diagnosis: The build failed with the error "Unable to parse HTML; parse5 error code unexpected-character-in-attribute-name" at line 70 in the file index.html, specifically at line 7 where there is an issue with the image tag attribute.

Solution: The error is caused by an extra quote in the img tag attribute in the index.html file. To fix this, remove the extra quote from the img tag's class attribute.

Replace:

<img class="avatar" "overlay" src="images/melinkedin.jpg">
With:

<img class="avatar overlay" src="images/melinkedin.jpg">
After making this change, commit the updated index.html file to your repository and trigger a new build.